% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-sent.R
\name{agrupa_tw_sent}
\alias{agrupa_tw_sent}
\title{Juntan informações de sentimento aos dados dos tweets}
\usage{
agrupa_tw_sent(tweets, tw_sent, path_arq = "data/tw_fim")
}
\arguments{
\item{tweets}{base limpa de tweets criada usando tidy_tweets}

\item{tw_sent}{base com valores de sentimento gerada por tw_sent_base}

\item{path_arq}{path e nome do arquivo rds que será salvo}
}
\value{

}
\description{
Juntan informações de sentimento aos dados dos tweets
}
