% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-sent.R
\name{prop_pos}
\alias{prop_pos}
\title{Cria gráfico com proporção de tweets positivos sobre quantidade de tweets negativos}
\usage{
prop_pos(tweets_fim, tema)
}
\arguments{
\item{tweets_fim}{base agrupada de tweets criada usando agrupa_tw_sent}

\item{tema}{string indicando palavrs(s)-chave utilizada(s) na coleta dos tweets}
}
\value{
gráfico ggplot com a proporção de tweets positivos sobre quantidade de tweets negativos por hora
}
\description{
Cria gráfico com proporção de tweets positivos sobre quantidade de tweets negativos
}
