% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-explore.R
\name{qtd_tw_tempo}
\alias{qtd_tw_tempo}
\title{Cria gráfico com o total de tweets por hora}
\usage{
qtd_tw_tempo(tw, tema)
}
\arguments{
\item{tw}{base limpa de tweets criada usando tidy_tweets}

\item{tema}{string com a(s) palavra(s)-chave utilizadas na coleta dos tweets}
}
\value{
gráfico ggplot
}
\description{
Cria gráfico com o total de tweets por hora
}
