% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-sent.R
\name{sent_hora}
\alias{sent_hora}
\title{Cria gráfico com média de sentimento por hora}
\usage{
sent_hora(tweets_fim, tema)
}
\arguments{
\item{tweets_fim}{base agrupada de tweets criada usando agrupa_tw_sent}

\item{tema}{string indicando palavrs(s)-chave utilizada(s) na coleta dos tweets}
}
\value{
gráfico ggplot com a média de sentimento por hora
}
\description{
Cria gráfico com média de sentimento por hora
}
