% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-import_tidy.R
\name{tidy_tweets}
\alias{tidy_tweets}
\title{Arruma base de tweets}
\usage{
tidy_tweets(arquivo_csv, path_tidy = "data")
}
\arguments{
\item{arquivo_csv}{Arquivo csv com metadados do Twitter}

\item{path_tidy}{Pasta em que será salvo o objeto rds com a base arrumada}
}
\value{
cria base tidy em data/
}
\description{
Arruma base de tweets
}
