% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-explore.R
\name{tw_hashtags}
\alias{tw_hashtags}
\title{Hashtags mais compartilhadas}
\usage{
tw_hashtags(tw)
}
\arguments{
\item{tw}{base limpa de tweets criada usando tidy_tweets}
}
\value{
df com 50 hashtags mais usadas e quantidade de vezes que foram compartilhadas
}
\description{
Hashtags mais compartilhadas
}
