% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-sent.R
\name{tw_pos_neg}
\alias{tw_pos_neg}
\title{Lista tweets mais positivos ou mais negativos}
\usage{
tw_pos_neg(tweets, tw_sent, sent = "positivo")
}
\arguments{
\item{tweets}{base limpa de tweets criada usando tidy_tweets}

\item{tw_sent}{base com valores de sentimento obtida por tw_sent_base}

\item{sent}{"positivo" ou "negativo"}
}
\value{
df com 10 tweets mais positivos/negativos
}
\description{
Lista tweets mais positivos ou mais negativos
}
