% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-sent.R
\name{tw_sent_base}
\alias{tw_sent_base}
\title{Calcula pontuação de sentimento de tweets}
\usage{
tw_sent_base(tw)
}
\arguments{
\item{tw}{base limpa de tweets criada usando tidy_tweets}
}
\value{
df com tweets e valores de sentimento
}
\description{
Calcula pontuação de sentimento de tweets
}
