% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-explore.R
\name{tw_urls}
\alias{tw_urls}
\title{URLs mais compartilhadas}
\usage{
tw_urls(tw)
}
\arguments{
\item{tw}{base limpa de tweets criada usando tidy_tweets}
}
\value{
df com 50 URLs mais citadas e quantidade de vezes que foram compartilhadas
}
\description{
URLs mais compartilhadas
}
